css overhaul
we were stacking up on text styles, and I thought it would be better to
switch to a convention more like bootstrap, where instead of having
separate styles each defining its own font, size, weight, color, etc.

We could just have classes that define a font, a size, a weight, a
color, and stack them on the elements. Discussed with Uri and he said,
“fine”.